+++ CHANGELOG.md
@@ -16,6 +16,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added a field to configure randomized backgrounds.
 - Added a config option to darken the background (closer matching the pre-1.12 achievements screen).
 
+### Fixed
+
+- Fixed screen shifting not working properly with the inventory button options.
+
 ## [3.0.1] - 2026-05-27
 
 ### Fixed